Session carries based on school/location. Read Description
If your IP matches a specific location, then it’s not necessary for you
to supply a location in the URL.
?s=…

Though since this script is being run on a server (IP not a campus IP),
a specific location is necessary and the cookie supplied to the script
needs to be carried over for every request… So that’s what this push
accomplishes. Carries over PHPSESS cookie ID so that a location is kept
consistent.